port tryMergingIntoMemset
we need to ensure that the newly inserted memset dominates all its uses.
Range.TheStores are ordered unpredictably because of the append operation in
MemsetRange::addRange (not ordered by BasicBlock traversal, sadly), so we can't
rely on that for memset positioning.

since the memset's only dependency is StartPtr, record explicitly the store node
that originally used this value, and insert adjacent to it.