clarify where non-locality is permitted
In the general case, suppose a transform T takes as input (I_0, I_1) and outputs
(I_0', I_1'). If I_0 == I_0' (I_0 is only read as a match for T), then it's
sufficient for I_0 to dominate I_1. Otherwise, I_1 must also post-dominate I_0.

For example, processMemCpyMemCpyDependence matches on

    memcpy(b <- a)
    memcpy(c <- b)

and outputs

    memcpy(b <- a)
    memcpy(c <- a)

Since it only modifies the second instruction and leaves the first unchanged,
it's enough for the first to dominate the second.

On the other hand, performCallSlotOptzn transforms

    f(a)
    memcpy(b <- a)

into f(b) and removes the memcpy. Thus, it's additionally necessary for the
memcpy to post-dominate the call to f.